+++ src/lib/handlers/externals/externals.contract.ts
@@ -4,7 +4,8 @@ import type { Version } from '../version/version.contract';
 
 type ExternalsHandler = {
     fromStorage(scope: string|'global'): Record<string, Version>;
-    toStorage(externals: SharedInfo[], scopeUrl: string): SharedInfo[]
+    toStorage(externals: SharedInfo[], scopeUrl: string): SharedInfo[];
+    checkForIncompatibleSingletons(externals: SharedInfo[]): void;
 }
 
 export { ExternalsHandler, SharedInfo, ShareObject, ShareConfig, ShareOptions }