+++ src/lib/handlers/version/version.contract.ts
@@ -1,15 +1,17 @@
 type Version = {
     version: string,
-    requiredVersion?: string,
-    strictRequiredVersion?: string,
+    requiredVersion?: [string,string],
+    strictRequiredVersion?: [string,string],
     url: string
 }
 
 type VersionHandler = {
     compareVersions: (v1: string, v2: string) => number
     getLatestVersion: (v1: Version, v2?: Version) => Version
-    isCompatible: (version: string, requiredVersion: string) => boolean,
-    stripVersionRange: (version: string) => string
+    isCompatible: (version: string, requiredVersion: [string, string]) => boolean,
+    toRange: (version: string) => [string,string],
+    toVersion: (requiredVersion: string) => string,
+    getSmallestVersionRange: (newRange: [string,string], currentRange?: [string,string]) => [string,string]
 }
 
 export {VersionHandler, Version}