+++ src/lib/handlers/version/version.handler.ts
@@ -9,18 +9,23 @@ const versionHandlerFactory = (): VersionHandler => {
         return SEMVER_REGEX.test(version);
     }
 
-    const toParts = (version:string): [number,number,number,string|undefined] => {
+    const toParts = (version:string): [number,number,number,string] => {
         if(!isSemver(version)) throw new NFError(`Invalid version '${version}'.`);
         const parts = [...version.matchAll(/^(\d+)\.(\d+)\.(\d+)(?:-([^+]+))?/g)][0]!;
-        return [Number(parts[1]), Number(parts[2]), Number(parts[3]), parts[4]?.toString()]
+        return [Number(parts[1]), Number(parts[2]), Number(parts[3]), !!parts[4] ? `-${parts[4]}` : ""]
     }
 
-    const stripVersionRange = (version:string) => (version.split(' ').pop() ?? version).replace(/^[~^>=<]+/, '');
+    const toVersion = (requiredVersion:string) => (requiredVersion.split(' ').pop() ?? requiredVersion).replace(/^[~^>=<]+/, '');
+
+    const getNextVersion = (version: string) => {
+        const [major, minor, patch] = toParts(version);
+        return `${major}.${minor}.${patch+1}`;
+    }
 
     const compareVersions = (v1: string, v2: string): number => {
         try{
-            const v1Parts = toParts(stripVersionRange(v1));
-            const v2Parts = toParts(stripVersionRange(v2));
+            const v1Parts = toParts(toVersion(v1));
+            const v2Parts = toParts(toVersion(v2));
     
             // Check numeric part (1.1.0)
             for (let i = 0; i < 3; i++) {
@@ -47,21 +52,43 @@ const versionHandlerFactory = (): VersionHandler => {
         }
     };
 
-    const isCompatible = (version: string, requiredVersion: string): boolean => {
-        const [major,minor,patch,_] = toParts(version);
-        
-
+    const toRange = (requiredVersion: string): [string,string] => {
         if(requiredVersion.startsWith("^")) {
-            const [reqMajor,reqMinor,reqPatch] = toParts(requiredVersion.slice(1));
-            return (major === reqMajor && (minor > reqMinor || (minor === reqMinor && patch >= reqPatch)))
+            const [major,minor,patch,flag] = toParts(requiredVersion.slice(1));
+            return [`${major}.${minor}.${patch}${flag}`, `${major+1}.0.0`]
         }
 
         if(requiredVersion.startsWith("~")) {
-            const [reqMajor,reqMinor,reqPatch] = toParts(requiredVersion.slice(1));
-            return (major === reqMajor && minor === reqMinor && patch >= reqPatch);
+            const [major,minor,patch,flag] = toParts(requiredVersion.slice(1));
+            return [`${major}.${minor}.${patch}${flag}`, `${major}.${minor+1}.0`]
         }
 
-        return version.localeCompare(requiredVersion) === 0;
+        const matchVersionRange = requiredVersion.match(`^(>=|>)([^\\s]+)\\s+(<=|<)([^\\s]+)$`);
+        if (matchVersionRange) {
+            const [_, minOperator, minVersion, maxOperator, maxVersion] = matchVersionRange;
+            if(!minVersion || !isSemver(minVersion)) throw new NFError(`Invalid min version '${requiredVersion}'`);
+            if(!maxVersion || !isSemver(maxVersion)) throw new NFError(`Invalid max version '${requiredVersion}'`);
+
+            const versionRange: [string,string] = [minVersion, maxVersion];
+
+            if(minOperator === ">") versionRange[0] = getNextVersion(minVersion);
+            if(maxOperator === "<=") versionRange[1] = getNextVersion(maxVersion);
+            
+            return versionRange;
+        }
+
+        if(isSemver(requiredVersion)) {
+            return [requiredVersion, getNextVersion(requiredVersion)];
+        }
+
+        throw new NFError(`Could not convert '${requiredVersion}' to a version range.`);
+    }
+
+
+    const isCompatible = (version: string, [minVersion, maxVersion]: [string,string]): boolean => {
+        const isAboveOrEqualToMin = compareVersions(version, minVersion) >= 0;    
+        const isBelowMax = compareVersions(version, maxVersion) < 0;    
+        return isAboveOrEqualToMin && isBelowMax;
     }
 
     const getLatestVersion = (newVersion: Version, currentVersion?: Version) => {
@@ -73,7 +100,14 @@ const versionHandlerFactory = (): VersionHandler => {
         return currentVersion;
     }
 
-    return {compareVersions, getLatestVersion, isCompatible, stripVersionRange};
+    const getSmallestVersionRange = (newRange: [string,string], currentRange?: [string,string]): [string,string] => {
+        if(!currentRange) return newRange;
+        const min = compareVersions(currentRange[0], newRange[0]) > 0 ? currentRange[0] : newRange[0];
+        const max = compareVersions(currentRange[1], newRange[1]) < 0 ? currentRange[1] : newRange[1];
+        return [min,max];
+    }
+
+    return {compareVersions, toRange, getLatestVersion, isCompatible, toVersion, getSmallestVersionRange};
 }
 
 export { versionHandlerFactory}